requirements.txt
Django==3.2.10
djangorestframework==3.14.0  
gunicorn==20.1.0